requirements.txt
flask
flask-sock
flask-cors
python-dotenv
websocket-client
PyJWT
toml